mortgage-application/ai-services/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10
PyMuPDF==1.23.8
pdf2image==1.17.0
pdfplumber==0.10.3
easyocr==1.7.0
numpy==1.25.2
opencv-python==4.8.1.78
requests==2.31.0
python-dotenv==1.0.0
